Extract out parameters tab. Refactor output pane.
