+++ x-pack/legacy/plugins/painless_playground/public/components/output_pane/output_tab.tsx
@@ -1,0 +1,27 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License;
+ * you may not use this file except in compliance with the Elastic License.
+ */
+
+import React from 'react';
+import { EuiCodeBlock, EuiSpacer } from '@elastic/eui';
+import { i18n } from '@kbn/i18n';
+
+import { formatResponse } from '../../lib/helpers';
+import { Response } from '../../common/types';
+
+interface Props {
+  response?: Response;
+}
+
+export function OutputTab({ response = {} }: Props) {
+  return (
+    <>
+      <EuiSpacer size="m" />
+      <EuiCodeBlock language="json" paddingSize="s" isCopyable>
+        {formatResponse(response.success || response.error)}
+      </EuiCodeBlock>
+    </>
+  );
+}
